- Pkg::ResolvableProperties() returns license and status if   available - fixed id of undefined repository (-1 returned as unsigned) - 2.15.40
svn path=/trunk/pkg-bindings/; revision=39920

+++ src/Callbacks.cc
@@ -513,7 +513,7 @@ namespace ZyppRecipients {
 	    size = pkg->downloadSize();
 
 	    // convert the repo ID
-	    int source_id = _pkg_ref.logFindAlias(pkg->repository().info().alias());
+	    long long source_id = _pkg_ref.logFindAlias(pkg->repository().info().alias());
 	    int media_nr = pkg->mediaNr();
 
 	    if( source_id != last_source_id || media_nr != last_source_media )
